+++ packages/core/src/StyleFilter.ts
@@ -1,18 +1,17 @@
-import Filter from './Filter';
+import Filter, { ElementAttributes } from './Filter';
 
 const INVALID_STYLES = /(url|image|image-set)\(/i;
 
 export default class StyleFilter extends Filter {
-  attribute(name: string, value: unknown): unknown {
-    if (name === 'style' && typeof value === 'object' && value) {
-      const styles = value as CSSStyleDeclaration;
-
-      Object.keys(styles).forEach(key => {
-        const prop = key as keyof CSSStyleDeclaration;
-
-        if (String(styles[prop]).match(INVALID_STYLES)) {
+  attribute<K extends keyof ElementAttributes>(
+    name: K,
+    value: ElementAttributes[K],
+  ): ElementAttributes[K] {
+    if (name === 'style') {
+      Object.keys(value).forEach(key => {
+        if (String(value[key]).match(INVALID_STYLES)) {
           // eslint-disable-next-line no-param-reassign
-          delete styles[prop];
+          delete value[key];
         }
       });
     }